IAtom.stereoParity: int -> Integer
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@8378 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/AtomTest.java
@@ -192,7 +192,7 @@ public class AtomTest extends CDKTestCase {
 
         IAtom a = builder.newAtom("C");
         a.setStereoParity(parity);
-        assertEquals(parity, a.getStereoParity());
+        assertEquals(parity, a.getStereoParity().intValue());
     }
     public void testGetStereoParity() {
         testSetStereoParity_int();
@@ -274,7 +274,7 @@ public class AtomTest extends CDKTestCase {
 
         // test cloning
         atom.setStereoParity(4);
-        assertEquals(3, clone.getStereoParity());
+        assertEquals(3, clone.getStereoParity().intValue());
     }
 
     /**
